move all other components into their own folders
