+++ client/components/audio-controls/audio-controls.test.tsx
@@ -4,7 +4,7 @@ import { beforeEach, describe, expect, it, vi } from 'vitest';
 import { KeyboardOptionsProvider } from '@/context/keyboard-options-context';
 import type { FullNote } from '@/types/keyboard-option-types';
 import { playSelectedNotes } from '@/utils/audio-control-functions';
-import AudioControls from './audio-controls';
+import AudioControls from '.';
 
 const mockFullNoteOptions: FullNote[] = [
 	'C4',
