+++ client/components/keyboard-full/index.tsx
