+++ client/components/keyboard-full/keyboard-full.test.tsx
@@ -2,7 +2,7 @@ import { render, screen } from '@testing-library/react';
 import { beforeEach, describe, expect, it } from 'vitest';
 
 import { KeyboardOptionsProvider } from '@/context/keyboard-options-context';
-import KeyboardFull from './keyboard-full';
+import KeyboardFull from '.';
 
 describe('Full Keyboard', () => {
 	beforeEach(() => {
