+++ client/components/keyboard-settings/keyboard-settings.test.tsx
@@ -2,7 +2,7 @@ import { render, screen } from '@testing-library/react';
 import { beforeEach, describe, expect, it } from 'vitest';
 
 import { KeyboardOptionsProvider } from '@/context/keyboard-options-context';
-import KeyboardSettings from './keyboard-settings';
+import KeyboardSettings from '.';
 
 describe('Keyboard settings', () => {
 	beforeEach(() => {
